Rename redis_repository to kebabcase
